moved search in archive logics in a separate component
